requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.9.3
certifi==2024.6.2
chardet==4.0.0
click==8.1.7
distro==1.9.0
Flask==2.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
openai==1.40.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dotenv==0.15.0
pytz==2024.1
redis==3.5.3
requests==2.25.1
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.5
twilio==6.45.4
typing_extensions==4.12.2
urllib3==1.26.18
Werkzeug==2.2.2